Modifying Code Alarms
issue:https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I649KL

+++ ecmascript/dfx/cpu_profiler/cpu_profiler.cpp
@@ -171,7 +171,7 @@ std::unique_ptr<struct ProfileInfo> CpuProfiler::StopCpuProfilerForInfo()
 
 void CpuProfiler::SetCpuSamplingInterval(int interval)
 {
-    interval_ = interval;
+    interval_ = static_cast<uint32_t>(interval);
 }
 
 void CpuProfiler::SetCallNapiGetStack(bool getStack)
